pyfinagent-app/requirements.txt
streamlit
requests
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-aiplatform
db-dtypes
pandas
fpdf2
Authlib>=1.3.2
plotly